reorganise duplicated resend code
